+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 ### Added
 - Technical - Add optional chaining plugin to babel.
+- Tests - Add MySQL min (5.6) and max (8.0) versions.
+- Database Support - Add support for MySQL 8.
 
 ### Changed
 - Technical - Upgrade to babel 7 stable.