+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Smart Actions - Add a method to `RecordsGetter` to get all models IDs given a query or an ID list.
 
 ## RELEASE 6.0.0-beta.0 - 2020-02-18
 ### Added
@@ -8,7 +10,7 @@
 - Tests - Add MySQL min (5.6) and max (8.0) versions.
 - Database Support - Add support for MySQL 8.
 - Resource Deletion - Users can now bulk delete records.
-- Smart Actions - Add a method to `RecordsGetter` to get all models IDs given a query or an ID list.
+- Technical - Add test and documentation for query-builder sort method.
 
 ### Changed
 - Technical - Upgrade to babel 7 stable.