+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - Smart Actions - Add a method to `RecordsGetter` to get all models IDs given a query or an ID list.
 
+### Changed
+- Initialisation - Make `Liana.init` return a promise resolving when all is up.
+
 ## RELEASE 6.0.0-beta.0 - 2020-02-18
 ### Added
 - Technical - Add optional chaining plugin to babel.
@@ -14,7 +17,6 @@
 
 ### Changed
 - Technical - Upgrade to babel 7 stable.
-- Initialisation - Make `Liana.init` return a promise resolving when all is up.
 
 ### Fixed
 - Technical - Remove cognitive complexity in query-builder service.